Change bID type definition to uint32_t

+++ simulatorCore/src/utils/tDefs.h
@@ -12,8 +12,8 @@
 #include <random>
 #include <functional>
 
-typedef uint64_t bID;           //!< Defines the type of a module identifier
-#define BID_MAX UINT64_MAX       //!< Maximum module identifier value
+typedef uint32_t bID;           //!< Defines the type of a module identifier
+#define BID_MAX UINT32_MAX       //!< Maximum module identifier value
 
 typedef uint64_t Time; //!< Time (unit: us)
 #define TIME_MAX UINT64_MAX //!< Maximum possible value of time given its type
